fixes typescript error
