+++ src/interfaces/authenticate.interface.tsx
@@ -1,0 +1,6 @@
+export interface IAuthenticateContext {
+  isAuthenticated: boolean;
+  accessToken: string | null;
+  setToken: (accessToken: string) => void;
+  clearToken: () => void;
+}
