+++ src/utils/useAuth.tsx
@@ -1,0 +1,4 @@
+import { useContext } from "react";
+import { AuthenticateContext } from "../store/authcontext";
+
+export const useAuth = () => useContext(AuthenticateContext);
